Added styling and secondary index page

+++ Staging/deleteRecord.php
@@ -13,30 +13,37 @@ if ($result->num_rows > 0) {
     // output data of each row
     while($row = $result->fetch_assoc()) {
         $updatingResidence=$row[$residenceChoice];
-        $updatedResidence= $updatingResidence+1;
-        echo "<br> $updatingResidence<br>";
-        echo $updatedResidence;
+        if ($updatingResidence>=5){
+            echo "Sorry there has been in error: database error $residenceChoice has too many spots left. Please contact your system administrator.<br>
+            < a href='mailto:somone@google.com'> This guy </a>";
+        }else{
+            $updatedResidence= $updatingResidence+1;
+            echo "<br> $updatingResidence<br>";
+            echo $updatedResidence;
+            $sql2= "UPDATE residence_halls SET $residenceChoice='$updatedResidence' WHERE $residenceChoice= $updatingResidence ";
+            $result2= $conn->query($sql2);
+            
+            if ($conn->query($sql2) === TRUE) {
+            echo "<br>Record updated successfully <br>";
+            } else {
+            echo "<br>Error updating record: " . $conn->error;
+            }
+            
+            $sql3 = "DELETE FROM residence_areas WHERE id = $id";
+            
+            if ($conn->query($sql3) === TRUE) {
+            echo "Record deleted successfully";
+            } else {
+            echo "Error deleting record: " . $conn->error;
+}
+        }
+        
 } 
 }else {
     echo "0 results";
 }
 
-$sql2= "UPDATE residence_halls SET $residenceChoice='$updatedResidence' WHERE $residenceChoice= $updatingResidence ";
-$result2= $conn->query($sql2);
 
-if ($conn->query($sql2) === TRUE) {
-    echo "<br>Record updated successfully <br>";
-} else {
-    echo "<br>Error updating record: " . $conn->error;
-}
-
-$sql3 = "DELETE FROM residence_areas WHERE id = $id";
-
-if ($conn->query($sql3) === TRUE) {
-    echo "Record deleted successfully";
-} else {
-    echo "Error deleting record: " . $conn->error;
-}
 
 
 $conn->close();
